re-styling download modal and input components
