+++ src/components/download-data/style.scss
@@ -45,6 +45,99 @@
 
 }
 
+#location-select {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    background-color: #F3F3F3;
+    margin-top: 1vh;
+    border-radius: 10px;
+    width: 302px;
+
+    .selection-mode {
+        margin-top: 2vh;
+    }
+
+    .input-container {
+        background-color: #F3F3F3;
+        height: 30px;
+    }
+
+    .location-list {
+        max-height: 200px;
+    }
+}
+
+.selection-mode {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-evenly;
+    align-items: center;
+    width: 300px;
+    height: 40px;
+    background-color: #FFFFFF;
+    border-radius: 100px;
+    border: 1px solid #B7B7B7;
+    box-shadow: 0px 0px 6px rgba(0, 0, 0, 0.25);
+    margin-top: 5vh;
+
+    .selected-option-p {
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        background-color: #DFFBFF;
+        border-radius: 100px;
+        padding-left: 25px;
+        padding-right: 25px;
+
+        .selected-option-text-p {
+            color: #ffffff;
+        }
+        img {
+            height: 20px;
+            width: 20px;
+        }
+    }
+    .selected-option-2-p {
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        background-color: #505050;
+        border-radius: 100px;
+        padding-left: 15px;
+        padding-right: 15px;
+
+        .selected-option-text-p {
+            color: #ffffff;
+        }
+        img {
+            height: 20px;
+            width: 20px;
+            margin-right: 10px;
+        }
+    }
+    .unselected-option-p {
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        padding-left: 15px;
+        padding-right: 15px;
+        box-shadow: 1px 1.5px 4px grey inset;
+        border-radius: 100px;
+        cursor: pointer;
+
+        .unselected-option-text-p {
+            color: #93949B;
+        }
+        img {
+            height: 20px;
+            width: 20px;
+            margin-right: 10px;
+        }
+    }
+}
+
 .download-modal {
     position: fixed;
     top: 50%;
@@ -57,6 +150,11 @@
     background-color: #ffffff;
     // TODO: replace box shadow with grayer background later
     box-shadow: 0 5px 15px rgba(0, 0, 0, 0.3); 
+
+    .selection-button {
+        width: 9vw;
+        background-color: #F3F3F3;
+    }
 }
 
 // remove unnecessary blue outline from modal (does not impact accessibility)
