+++ src/components/input-components/style.scss
@@ -1,10 +1,11 @@
 .selection-button {
     padding-left: 20px;
-    background: #EDEDEE;
     height: 43px;
     font-family: 'Inter';
     font-size: 18px; 
     border: none;
     cursor: pointer;
     letter-spacing: 1.5px;
+    border-radius: 10px;
+    margin-right: 5px;
 }
