+++ src/screens/prediction/components/selection-bar/style.scss
@@ -28,6 +28,10 @@
             display: flex;
             flex-direction: row;
 
+            .selection-button {
+                width: 9vw;
+            }
+
             .predictionbar-year-selection-option {
                 font-size: 16px;
             }
