+++ src/screens/trapping-data/components/selection-bar/style.scss
@@ -28,6 +28,10 @@
             display: flex;
             flex-direction: row;
 
+            .selection-button {
+                width: 9vw;
+            }
+
             .historicalbar-year-selection-option {
                 font-size: 16px;
             }
